refactor: extract data models to separate module
